+++ libclc/amdgcn/lib/workitem/get_local_id.cl
@@ -1,6 +1,6 @@
 #include <clc/clc.h>
 
-_CLC_DEF uint get_local_id(uint dim)
+_CLC_DEF size_t get_local_id(uint dim)
 {
 	switch(dim) {
 	case 0: return __builtin_amdgcn_workitem_id_x();
